consumer/requirements.txt
kafka-python
pymongo
prometheus_client
